Move not used functions into Hydra.Chain.Direct.State
